+++ shell.nix
@@ -1,14 +1,14 @@
-{}:
-let
-  # Pin the deployment package-set to a specific version of nixpkgs
-  pkgs = import
+{ pkgs ? let
+    lockfile = builtins.fromJSON (builtins.readFile ./flake.lock);
+    nixpkgs = lockfile.nodes.nixpkgs.locked;
+  in
+  import
     (builtins.fetchTarball {
-      url = "https://github.com/NixOS/nixpkgs/archive/2f9fd351ec37f5d479556cd48be4ca340da59b8f.tar.gz";
-      sha256 = "0w3ysrhbqhgr1qnh0r9miyqd7yf7vsd4wcd21dffwjlb99lynla8";
+      url = "https://github.com/NixOS/nixpkgs/archive/${nixpkgs.rev}.tar.gz";
+      sha256 = nixpkgs.narHash;
     })
-    { };
-
-in
+    { }
+}:
 with pkgs;
 
 stdenv.mkDerivation {