+++ CHANGELOG.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Nix Flake
+- Docker image build using Nix `dockerTools`
+
 ### Changed
 
 - migrated to ESM project
 - format of a config file changed from .ts to .yaml (due to ESM migration)
 - regenerated yarn.lock
-- updated nixpkgs and hence NodeJS to 16.18.1
+- nixpkgs updated to latest `nixos-22.11` and hence NodeJS to `16.18.1`
 
 ## [1.4.0] - 2023-02-07